+++ src/runtime/editor-action-access.ts
@@ -363,8 +363,14 @@ export class EditorActionAccessFactory {
             updateMaskList: () => {
                 callbacks.updateMaskList();
             },
+            updateMaskListSelection: (mask) => {
+                callbacks.updateMaskListSelection(mask);
+            },
             updateAnnotationList: () => {
                 callbacks.updateAnnotationList();
+            },
+            updateAnnotationListSelection: (annotation) => {
+                callbacks.updateAnnotationListSelection(annotation);
             },
             updateUi: () => {
                 callbacks.updateUi();